\name{SpecDataEgg}
\alias{SpecDataEgg}
\alias{MetaDataEgg}
\docType{data}
\title{Package data.}
\description{Metadata and spectral data for SpecLibDolphinBlackSkimmerEgg2014. Called by \code{LibraryReport()}.}
\usage{SpecDataEgg}
\format{\code{SpecDataEgg} is a data frame containing the mass spectra.  \code{MetaDataEgg} is a data frame containing the metadata.
}
\keyword{datasets} 

