\name{SpecLibBlackSkimmerEgg2014-package}
\alias{SpecLibBlackSkimmerEgg2014-package}
\docType{package}
\title{
Mass Spectral Library San Diego Bay Black Skimmer Bird Eggs 2014
}
\description{
Contains the code and mass spectral library described in: Identifying Bioaccumulative Halogenated Organic Compounds Using a Nontargeted Analytical Approach: Seabirds as Sentinels.

The corresponding article includes a description of
the study goals, samples, and GCxGC/TOF-MS
analytical method. The original PDF version of this library is supplementary
material to the article.
}
\keyword{ package }
